Restructure biometricsClient:
- Create package
- Create a file by model
