+++ app/src/main/java/org/dhis2/data/biometrics/biometricsClient/models/SimprintsIdentifiedItem.kt
@@ -1,0 +1,8 @@
+package org.dhis2.data.biometrics.biometricsClient.models
+
+data class SimprintsIdentifiedItem(
+    val guid: String,
+    val confidence: Float,
+    val isLinkedToCredential: Boolean,
+    val isVerified: Boolean?,
+)
