+++ app/src/main/java/org/dhis2/data/biometrics/biometricsClient/models/SimprintsRegisteredItem.kt
@@ -1,0 +1,12 @@
+package org.dhis2.data.biometrics.biometricsClient.models
+
+data class SimprintsRegisteredItem(
+    val guid: String,
+    val hasCredential: Boolean,
+    val scannedCredential: BiometricsCredential?
+)
+
+data class BiometricsCredential(
+    val credentialType: String,
+    val value: String,
+)
