+++ app/src/main/java/org/dhis2/data/biometrics/biometricsClient/models/VerifyResult.kt
@@ -1,0 +1,8 @@
+package org.dhis2.data.biometrics.biometricsClient.models
+
+sealed class VerifyResult {
+    data object Match : VerifyResult()
+    data object NoMatch : VerifyResult()
+    data object Failure : VerifyResult()
+    data object AgeGroupNotSupported : VerifyResult()
+}
