+++ app/src/main/java/org/dhis2/usescases/biometrics/duplicates/BiometricsDuplicatesDialogPresenter.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.map
 import org.dhis2.commons.filters.FilterManager
 import org.dhis2.commons.prefs.BasicPreferenceProvider
 import org.dhis2.commons.schedulers.SchedulerProvider
-import org.dhis2.data.biometrics.SimprintsIdentifiedItem
+import org.dhis2.data.biometrics.biometricsClient.models.SimprintsIdentifiedItem
 import org.dhis2.data.search.SearchParametersModel
 import org.dhis2.usescases.searchTrackEntity.SearchRepository
 import org.dhis2.usescases.searchTrackEntity.SearchRepositoryKt
