+++ app/src/main/java/org/dhis2/usescases/biometrics/duplicates/LastPossibleDuplicates.kt
@@ -1,6 +1,6 @@
 package org.dhis2.usescases.biometrics.duplicates
 
-import org.dhis2.data.biometrics.SimprintsIdentifiedItem
+import org.dhis2.data.biometrics.biometricsClient.models.SimprintsIdentifiedItem
 
 data class LastPossibleDuplicates(
     val guids: List<SimprintsIdentifiedItem>,
