+++ app/src/test/java/org/dhis2/usescases/biometrics/GetOtgUnitAsModuleIdByBasicInfo.kt
@@ -1,6 +1,6 @@
 package org.dhis2.usescases.biometrics
 
-import org.dhis2.data.biometrics.BiometricsClient
+import org.dhis2.data.biometrics.biometricsClient.BiometricsClient
 import org.junit.Assert.assertEquals
 import org.junit.Test
